PSYCH doing it rn (shoutout refactoring)
