+++ src/atomate2/jdftx/io/JDFTXOutfile.py
@@ -1,6 +1,9 @@
 import os
 from functools import wraps
 import math
+from ase import Atom, Atoms
+from ase.units import Bohr
+from jdftx.io.out_to_log import get_start_line
 import numpy as np
 from dataclasses import dataclass, field
 import scipy.constants as const
@@ -423,13 +426,6 @@ class JDFTXOutfile(ClassPrintFormatter):
         # if text[line].split()[1] != 'no':
         #     raise ValueError('kpoint-reduce-inversion must = no in single point DFT runs so kgrid without time-reversal symmetry is used (BGW requirement)')
 
-        # eigstats = cls._get_eigstats_varsdict(text, instance.prefix)
-        # instance.Emin = eigstats["Emin"]
-        # instance.HOMO = eigstats["HOMO"]
-        # instance.EFermi = eigstats["EFermi"]
-        # instance.LUMO = eigstats["LUMO"]
-        # instance.Emax = eigstats["Emax"]
-        # instance.Egap = eigstats["Egap"]
         instance._set_eigvars(text)
         print(f"Egap: {instance.Egap}")
         if instance.broadening_type is not None:
@@ -627,3 +623,214 @@ class JDFTXOutfile(ClassPrintFormatter):
             value = getattr(self, field)
             dct[field] = value
         return dct
+
+
+def get_input_coord_vars_from_outfile(outfname):
+    start_line = get_start_line(outfname)
+    names = []
+    posns = []
+    R = np.zeros([3,3])
+    lat_row = 0
+    active_lattice = False
+    with open(outfname) as f:
+        for i, line in enumerate(f):
+            if i > start_line:
+                tokens = line.split()
+                if len(tokens) > 0:
+                    if tokens[0] == "ion":
+                        names.append(tokens[1])
+                        posns.append(np.array([float(tokens[2]), float(tokens[3]), float(tokens[4])]))
+                    elif tokens[0] == "lattice":
+                        active_lattice = True
+                    elif active_lattice:
+                        if lat_row < 3:
+                            R[lat_row, :] = [float(x) for x in tokens[:3]]
+                            lat_row += 1
+                        else:
+                            active_lattice = False
+                    elif "Initializing the Grid" in line:
+                        break
+    if not len(names) > 0:
+        raise ValueError("No ion names found")
+    if len(names) != len(posns):
+        raise ValueError("Unequal ion positions/names found")
+    if np.sum(R) == 0:
+        raise ValueError("No lattice matrix found")
+    return names, posns, R
+
+
+def get_atoms_from_outfile_data(names, posns, R, charges=None, E=0, momenta=None):
+    atoms = Atoms()
+    posns *= Bohr
+    R = R.T*Bohr
+    atoms.cell = R
+    if charges is None:
+        charges = np.zeros(len(names))
+    if momenta is None:
+        momenta = np.zeros([len(names), 3])
+    for i in range(len(names)):
+        atoms.append(Atom(names[i], posns[i], charge=charges[i], momentum=momenta[i]))
+    atoms.E = E
+    return atoms
+
+
+def get_atoms_list_from_out_reset_vars(nAtoms=100, _def=100):
+    R = np.zeros([3, 3])
+    posns = []
+    names = []
+    chargeDir = {}
+    active_lattice = False
+    lat_row = 0
+    active_posns = False
+    log_vars = False
+    coords = None
+    new_posn = False
+    active_lowdin = False
+    idxMap = {}
+    j = 0
+    E = 0
+    if nAtoms is None:
+        nAtoms = _def
+    charges = np.zeros(nAtoms, dtype=float)
+    forces = []
+    active_forces = False
+    coords_forces = None
+    return R, posns, names, chargeDir, active_posns, active_lowdin, active_lattice, posns, coords, idxMap, j, lat_row, \
+        new_posn, log_vars, E, charges, forces, active_forces, coords_forces
+
+
+def get_atoms_list_from_out_slice(outfile, i_start, i_end):
+    charge_key = "oxidation-state"
+    opts = []
+    nAtoms = None
+    R, posns, names, chargeDir, active_posns, active_lowdin, active_lattice, posns, coords, idxMap, j, lat_row, \
+        new_posn, log_vars, E, charges, forces, active_forces, coords_forces = get_atoms_list_from_out_reset_vars()
+    for i, line in enumerate(open(outfile)):
+        if i > i_start and i < i_end:
+            if new_posn:
+                if "Lowdin population analysis " in line:
+                    active_lowdin = True
+                elif "R =" in line:
+                    active_lattice = True
+                elif "# Forces in" in line:
+                    active_forces = True
+                    coords_forces = line.split()[3]
+                elif line.find('# Ionic positions in') >= 0:
+                    coords = line.split()[4]
+                    active_posns = True
+                elif active_lattice:
+                    if lat_row < 3:
+                        R[lat_row, :] = [float(x) for x in line.split()[1:-1]]
+                        lat_row += 1
+                    else:
+                        active_lattice = False
+                        lat_row = 0
+                elif active_posns:
+                    tokens = line.split()
+                    if len(tokens) and tokens[0] == 'ion':
+                        names.append(tokens[1])
+                        posns.append(np.array([float(tokens[2]), float(tokens[3]), float(tokens[4])]))
+                        if tokens[1] not in idxMap:
+                            idxMap[tokens[1]] = []
+                        idxMap[tokens[1]].append(j)
+                        j += 1
+                    else:
+                        posns = np.array(posns)
+                        active_posns = False
+                        nAtoms = len(names)
+                        if len(charges) < nAtoms:
+                            charges = np.zeros(nAtoms)
+                ##########
+                elif active_forces:
+                    tokens = line.split()
+                    if len(tokens) and tokens[0] == 'force':
+                        forces.append(np.array([float(tokens[2]), float(tokens[3]), float(tokens[4])]))
+                    else:
+                        forces = np.array(forces)
+                        active_forces = False
+                ##########
+                elif "Minimize: Iter:" in line:
+                    if "F: " in line:
+                        E = float(line[line.index("F: "):].split(' ')[1])
+                    elif "G: " in line:
+                        E = float(line[line.index("G: "):].split(' ')[1])
+                elif active_lowdin:
+                    if charge_key in line:
+                        look = line.rstrip('\n')[line.index(charge_key):].split(' ')
+                        symbol = str(look[1])
+                        line_charges = [float(val) for val in look[2:]]
+                        chargeDir[symbol] = line_charges
+                        for atom in list(chargeDir.keys()):
+                            for k, idx in enumerate(idxMap[atom]):
+                                charges[idx] += chargeDir[atom][k]
+                    elif "#" not in line:
+                        active_lowdin = False
+                        log_vars = True
+                elif log_vars:
+                    if np.sum(R) == 0.0:
+                        R = get_input_coord_vars_from_outfile(outfile)[2]
+                    if coords != 'cartesian':
+                        posns = np.dot(posns, R)
+                    if len(forces) == 0:
+                        forces = np.zeros([nAtoms, 3])
+                    if coords_forces.lower() != 'cartesian':
+                        forces = np.dot(forces, R)
+                    opts.append(get_atoms_from_outfile_data(names, posns, R, charges=charges, E=E, momenta=forces))
+                    R, posns, names, chargeDir, active_posns, active_lowdin, active_lattice, posns, coords, idxMap, j, lat_row, \
+                        new_posn, log_vars, E, charges, forces, active_forces, coords_forces = get_atoms_list_from_out_reset_vars(
+                        nAtoms=nAtoms)
+            elif "Computing DFT-D3 correction:" in line:
+                new_posn = True
+    return opts
+
+
+def get_start_lines(outfname, add_end=False):
+    start_lines = []
+    end_line = 0
+    for i, line in enumerate(open(outfname)):
+        if "JDFTx 1." in line:
+            start_lines.append(i)
+        end_line = i
+    if add_end:
+        start_lines.append(i)
+    return start_lines
+
+
+def get_atoms_list_from_out(outfile):
+    start_lines = get_start_lines(outfile, add_end=True)
+    atoms_list = []
+    for i in range(len(start_lines) - 1):
+        atoms_list += get_atoms_list_from_out_slice(outfile, start_lines[i], start_lines[i+1])
+    return atoms_list
+
+
+def is_done(outfile):
+    start_line = get_start_line(outfile)
+    done = False
+    with open(outfile, "r") as f:
+        for i, line in enumerate(f):
+            if i > start_line:
+                if "Minimize: Iter:" in line:
+                    done = False
+                elif "Minimize: Converged" in line:
+                    done = True
+    return done
+
+
+def get_initial_lattice(outfile, start):
+    start_key = "lattice  \\"
+    active = False
+    R = np.zeros([3, 3])
+    lat_row = 0
+    for i, line in enumerate(open(outfile)):
+        if i > start:
+            if active:
+                if lat_row < 3:
+                    R[lat_row, :] = [float(x) for x in line.split()[0:3]]
+                    lat_row += 1
+                else:
+                    active = False
+                    lat_row = 0
+            elif start_key in line:
+                active = True
+    return R
